requirements.txt
  # build
  cmake
  ruamel.yaml
  # lint
  pre-commit
  # testing
  build==0.7.0
  hypothesis
  pexpect
  pytest
  pytest-asyncio
  pytest-cov
  pytest-httpserver
  pytest-instafail
  pytest-rerunfailures
  pytest-xdist
  selenium==4.1.0
  # maintenance
  bump2version
